\name{cliHelp}
\alias{cliHelp}
\title{
Help / Usage
}
\description{
Provides help and usage information from the depthmapXcli.
}
\usage{
cliHelp(cliPath = getDefaultCLILocation())
}
\arguments{
  \item{cliPath}{
    Optional. Provide the path to a different CLI.
  }
}
\value{
    Prints the output from the depthmapXcli.
}
\author{
    Petros Koutsolampros
}
